asr_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.26.2
pydub==0.25.1
python-dotenv==1.0.0
aiofiles==23.2.1
transformers==4.35.2
accelerate==0.25.0
